fix(PUPIL-269): move checkbox border styles to the outline
this ensures that the `border-radius` is applied inside the border
(as it is in Figma). This also ensures that the border does not subtract
from the width and height of the container so the layout is not impacted.

The other option is to switch to `content-box` and play with negative
margins which no one wants to do.