feat(PUPIL-229): add `OakLessonTopNav` component
This will render the navigation bar at the top of a lesson.

It includes slots for the back link and counter. These are included
to keep the component dumb and leave it up to the host to decide how to
hook up the back link and counter to the lesson state.

A `mobileSummary` prop is included to allow the host to pass in alternate
content when at the mobile breakpoint since the counter is not displayed.