feat(PUPIL-240)!: serve images directly from Cloudinary when possible
BREAKING CHANGE: Introduces a new environment variable `NEXT_PUBLIC_CLOUDINARY_CLOUD_NAME`
which is required by `next-cloudinary`.

until now images were always proxied by the Next server  where they are
resized on the fly. There is a considerable delay while the web-app pulls the full-size image
from Cloudinary before it can resize and serve it. This is not necessary for
images that are already in Cloudinary since that is exactly what Cloudinary is for.