fix: `OakCookieSettingsModal` being stacked behind `OakCookieBanner`
the z-index of 99999 caused that the banner to be rendered on top
of the settings modal. This commit restores the original values while adding
an option to set a stacking context for the entire `OakCookieConsent` component.

This ensures that the entire component can be stacked according to the needs of the
host app.